Split connection.rs
